src/main.py


import os
from mistralai.client import MistralClient
from mistralai.models.chat_completion import ChatMessage
from dotenv import load_dotenv

load_dotenv()

api_key = os.environ["MISTRAL_API_KEY"]
model = "mistral-large-latest"

client = MistralClient(api_key=api_key)

chat_response = client.chat(
    model=model,
    messages=[ChatMessage(role="user", content="Hello there!")]
)

print(chat_response.choices[0].message.content)


# TODO
# Initialize the strategy

# Launch the game

# Iterate over the game