requirements.txt
annotated-types==0.7.0
anyio==4.3.0
certifi==2024.2.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mistralai==0.2.0
orjson==3.10.3
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.0
selenium==4.21.0
python-dotenv==1.0.1